Made max/min limits for the blob counter larger (need for 1920x1200)
Added and max / min check for the limit of the rectangles we get, this will exclude the bottom hero selected abilities

Added some percentile math to get all the rectangles are are close to each other. The old code was including some 1/2 ult pictures, these then were not matching correctly. I.e 1/2 Wind Ranger was matching to Kunkka